Refactor native OpenMPI deployments to share between lammps/ kernels
