+++ k8s/namespace.yml.j2
@@ -1,0 +1,6 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: {{ native_mpi_namespace }}
+  labels:
+    name: {{ native_mpi_namespace }}
