+++ k8s/templated/namespace-kernels.yml
@@ -1,0 +1,6 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: openmpi-kernels
+  labels:
+    name: openmpi-kernels
