+++ k8s/templated/namespace-lammps.yml
@@ -1,6 +1,6 @@
 apiVersion: v1
 kind: Namespace
 metadata:
-  name: faasm-openmpi
+  name: openmpi-lammps
   labels:
-    name: faasm-openmpi
+    name: openmpi-lammps
