+++ tasks/kernels/__init__.py
@@ -2,5 +2,6 @@ from invoke import Collection
 
 from . import build
 from . import container
+from . import native
 
-ns = Collection(build, container)
+ns = Collection(build, container, native)
